requirements.txt
blurhash==1.1.4
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
charset-normalizer==2.0.10; python_full_version >= "3.6.0" and python_version >= "3"
decorator==5.1.1; python_version >= "3.5"
idna==3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
mastodon.py==1.5.1
pytelegrambotapi==4.3.1
python-dateutil==2.8.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
python-dotenv==0.19.2; python_version >= "3.5"
python-magic==0.4.24; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pytz==2021.3
requests==2.27.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
urllib3==1.26.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
